exported snackbar to RootLayout, added redirect to login, and previous path to state
